chore: procutil move conversion function to a separate file
